consolidated round robin and chunking, accepting input parameters
